Optimized Visualization
Fixed Issue that when robot point at tablet the tracker failed. This is fixed by making the robot look at the original location of the target. However, this still needs optimization that making the robot recognize the new target as the original target. A learning process is needed.  
Found problem that when initializing SoundFeedback Class, the robot will stand up.